Significant refactoring of input utilities
